+++ live/tests/acceptance/l1-signaler-une-epreuve-test.js
@@ -1,0 +1,67 @@
+import { describe, it, before, after } from 'mocha';
+import { expect } from 'chai';
+import startApp from '../helpers/start-app';
+import destroyApp from '../helpers/destroy-app';
+
+const FEEDBACK_FORM = '.feedback-panel__form';
+
+describe('Acceptance | Signaler une épreuve', function() {
+
+  let application;
+
+  before(function() {
+    application = startApp();
+  });
+
+  after(function() {
+    destroyApp(application);
+  });
+
+  function assertThatFeedbackPanelExist() {
+    expect(find('.feedback-panel')).to.have.lengthOf(1);
+  }
+
+  function assertThatFeedbackFormIsClosed() {
+    expect(find('.feedback-panel__open-link')).to.have.lengthOf(1);
+    expect(find(FEEDBACK_FORM)).to.have.lengthOf(0);
+  }
+
+  function assertThatFeedbackFormIsOpen() {
+    expect(find('.feedback-panel__open-link')).to.have.lengthOf(0);
+    expect(find(FEEDBACK_FORM)).to.have.lengthOf(1);
+  }
+
+  it('l1.1 Je peux signaler une épreuve depuis la page de comparaison de mes réponses avec celles attendues pour une épreuve (page de résultat du test)', async () => {
+    await visit('/assessments/ref_assessment_id/results/compare/ref_answer_qcm_id/1');
+    assertThatFeedbackPanelExist();
+  });
+
+  it('l1.2 Je peux signaler une épreuve directement depuis l\'épreuve en question', async () => {
+    await visit('/assessments/ref_assessment_id/challenges/ref_qcu_challenge_id');
+    assertThatFeedbackPanelExist();
+  });
+
+  it('l1.3 Le formulaire de signalement d\'une épreuve est remis à zéro dès que je change d\'épreuve', async () => {
+    await visit('/assessments/ref_assessment_id/challenges/ref_qru_challenge_id');
+    assertThatFeedbackFormIsClosed();
+
+    await click('.feedback-panel__open-link');
+    assertThatFeedbackFormIsOpen();
+
+    await click('.challenge-actions__action-skip');
+    assertThatFeedbackFormIsClosed();
+  });
+
+  it('l1.4 Le formulaire de signalement est remis à zéro même quand les 2 épreuves qui s\'enchaînent utilisent le même composant challenge-item-* (ex : q1 est de type "QCU" et q2 "QRU" ; toutes deux utilisent le composant challenge-item-qcu)', async () => {
+    // In our Mirage data set, in the "ref course", the QCU challenge is followed by a QRU's one
+    await visit('/assessments/ref_assessment_id/challenges/ref_qcu_challenge_id');
+    assertThatFeedbackFormIsClosed();
+
+    await click('.feedback-panel__open-link');
+    assertThatFeedbackFormIsOpen();
+
+    await click('.challenge-actions__action-skip');
+    assertThatFeedbackFormIsClosed();
+  });
+
+});